+++ neuroscheme/layouts/CameraBasedLayout.cpp
@@ -75,7 +75,10 @@ namespace neuroscheme
       }
       else
       {
-        auto graphicsItem = graphicsItemRep->item( _scene );
+        auto graphicsItem = graphicsItemRep->item( &_canvas->scene( ));
+        if ( graphicsItem->parentItem( ))
+          continue;
+
         auto item = dynamic_cast< Item* >( graphicsItem );
         auto obj = dynamic_cast< QObject* >( graphicsItem );
         if ( graphicsItem && item )