+++ Changelog.md
@@ -1,6 +1,7 @@
 # NeuroScheme
 
 ## git master
+* [!6] Added home button. Added circular layout. Fixed bugs.
 * [!5] Added scoop for color mappers
 * [!4] Enabled horizontal and vertical splitting of panes
 * First version of the refactor of legacy neuroscheme